Release 6.5.0

+++ upgrade/Upgrade-4.2.0.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * Since 2007 PayPal
  *
@@ -57,7 +58,7 @@ function upgrade_module_4_2_0($module)
     if (Configuration::get('PAYPAL_METHOD') == 'EXPRESS_CHECKOUT') {
         if (Configuration::get('PAYPAL_LIVE_ACCESS') || Configuration::get('PAYPAL_SANDBOX_ACCESS')) {
             Configuration::updateValue('PAYPAL_METHOD', 'EC');
-            if (Country::getIsoById(Configuration::get('PS_COUNTRY_DEFAULT')) == 'FR' || Country::getIsoById(Configuration::get('PS_COUNTRY_DEFAULT')) == 'UK') {
+            if (Country::getIsoById((int) Configuration::get('PS_COUNTRY_DEFAULT')) == 'FR' || Country::getIsoById((int) Configuration::get('PS_COUNTRY_DEFAULT')) == 'UK') {
                 Configuration::updateValue('PAYPAL_API_CARD', 0);
             }
         } else {
